Replaced AlignmentTransducer with AlignmentHypergraph throughout.
Portions of unit tests involving max-margin objectives are failing. Can we
attribute this to breaking ties between optimal alignments? It is just a
coincidence that the test involving printAlignment() passes? I suppose the
{stress, sutoressu} pair has a unique max-scoring alignment? Note that the
objective values returned by optimizers are higher under the new regime.
Maybe inspect the output of the opimizer to see what's going on?

+++ tests/TestStringEditHypergraph.cpp
@@ -16,6 +16,7 @@
 #include <math.h>
 #include <string>
 #include <vector>
+
 using namespace boost;
 
 BOOST_AUTO_TEST_CASE(testStringEditHypergraph)
@@ -110,11 +111,11 @@ BOOST_AUTO_TEST_CASE(testStringEditHypergraph)
   
   // Check that the Viterbi score is correct.
   double viterbiScore = model->viterbiScore(W, *pair, label);
-  BOOST_CHECK_CLOSE((double)viterbiScore, -300, 1e-8);
+  BOOST_CHECK_CLOSE(viterbiScore, -300, 1e-8);
   
   SparseRealVec realFv(d);
   double maxScore = model->maxFeatures(W, realFv, *pair, label, true);
-  BOOST_CHECK_CLOSE((double)maxScore, (double)viterbiScore, 1e-8);
+  BOOST_CHECK_CLOSE(maxScore, viterbiScore, 1e-8);
   
   // Check that values in the max-scoring feature vector are correct.  
   BOOST_CHECK_CLOSE((double)realFv[iIns], 3, 1e-4);
